Вносит правки по рекомендациям наставника
