+++ src/api.js
@@ -1,25 +1,25 @@
 import axios from 'axios';
+import {TIMER, Error, URL} from './utils/consts';
 
-const Error = {
-  UNAUTHORIZED: 401
-};
 
 export const createAPI = (onUnauthorized) => {
   const api = axios.create({
-    baseURL: `https://4.react.pages.academy/wtw`,
-    timeout: 1000 * 5,
+    baseURL: URL,
+    timeout: TIMER,
     withCredentials: true,
   });
 
-  const onSuccess = (response) => {
-    return response;
-  };
+  const onSuccess = (response) => response;
 
   const onFail = (err) => {
     const {response} = err;
 
     if (response.status === Error.UNAUTHORIZED) {
       onUnauthorized();
+
+      // Бросаем ошибку, потому что нам важно прервать цепочку промисов после запроса авторизации.
+      // Запрос авторизации - это особый случай и важно дать понять приложению,
+      // что запрос был неудачным.
       throw err;
     }
