+++ src/store/user/selectors.js
@@ -1,4 +1,8 @@
 import NameSpace from '../name-space';
 
+const getUserState = (state) => state[NameSpace.USER];
 
-export const getAuthorizationStatus = (store) => store[NameSpace.USER].authorizationStatus;
+export const getAuthorizationStatus = (state) => getUserState(state).authorizationStatus;
+export const getIsAuthorizing = (state) => getUserState(state).isAuthorizing;
+export const getIsAuthorizationError = (state) => getUserState(state).isAuthorizationError;
+export const getUserData = (state) => getUserState(state).userData;
