+++ src/utils/consts.js
@@ -55,4 +55,13 @@ export const EntryPoint = {
   FAVORITE: `/favorite`,
   MOVIES: `/films`,
   PROMO: `/films/promo`,
+  LOGIN: `/login`,
 };
+
+export const TIMER = 5000;
+
+export const Error = {
+  UNAUTHORIZED: 401,
+};
+
+export const URL = `https://4.react.pages.academy/wtw`;
